Replace Bash terminology with math terms
When I wrote this program, I took inspiration from shell and
pseudo-terminal programs such as Bash and Nodejs. Thus, when I came up
with variable names and error messages, I used Bash terminology such as
"command" and "arguments." This commit replaces the Bash terminology
with math terms. Doing so makes it easier for non-programmers to
understand the program.